refactor: move nested-key.d.ts
