+++ core/static/bundled/types/nested-key.d.ts
